Walls Removed When Maze Algorithm is Started
The Walls get removed when the Algorithm is started. However, the maze does not get completely generated, the "cursor" that is generating the maze "gets stuck". I will try to fix this problem in the next update. I used this video to help me: https://www.youtube.com/watch?v=8Ju_uxJ9v44